src/python/requirements.txt
langchain-openai>=0.3.0
azure-identity>=1.15.0
python-dotenv>=1.2.1
